Move not needed code from .h to .cpp
